重構影像/畫布與 GUI 繪製 API
- 重新整理 BITVIEW / BITMAP / BITVIEW_MEM 結構，加入旋轉欄位與多位元組像素支援（1/2/3/4 bytes-per-pixel），合併並簡化 Dot/DotPoint 讀寫介面。
- 新增 MEMDEV（繼承 LCDDEV + BITVIEW_MEM）作為離屏繪圖裝置；Window 離屏流程改為使用 MemDev 並以 pDev->BitsPerPixel() 分配緩衝，再 blit 回主畫面。
- 將虛擬函式 FillRect 改名為 rFill（包含簽名與所有呼叫端更新）以統一語意並避免混淆。
- LCDDEV::SetBitmap 改為接受 palette 與 HasTrans 參數（PCLOGPALETTE pPalEntries = nullptr, bool HasTrans = false）；GUI_DrawBitmap 簽名改為 GUI_DrawBitmap(CBITMAP& bm, POINT Pos)；呼叫流程改為先建立 BITVIEW（bm.At(Pos)）再呼 SetBitmap。
- 重寫 BITVIEW_MEM 的記憶體配置與 Alloc()，自動計算 BytesPerLine 與 pData，並支援 RGB24/32 的像素存取。
- 擴充文字對齊：新增 TEXTALIGN_CENTER（水平+垂直置中）。重寫 GUI_DispStringInRect* 系列以支援更完整的水平/垂直置中、分行與行數計算；移除/合併舊有的字串/字元繪製函式。
- 多個 widget（ListBox、Menu、ProgBar、DropDown、Radio、ScrollBar、Slider 等）做 API/成員重命名、const-correctness 調整，改為使用 rFill 與 GUI_DispStringInRect 做文字與背景繪製；ListBox Item 大小快取改為 mutable 以維持 const。
- OwnerDraw 回呼型別改為接受 const Widget*，並修正所有回呼與實作以符合 const-correctness。
- 若干平台/範例程式（包括 GUI_X、main.cpp、SimClient）更新為新介面（rFill、GUI_DrawBitmap(*pBm)、SetOnKey 強制轉型等）。

影響：大範圍 API/欄位與簽名變更，合併此提交需同步更新所有使用相關介面的程式碼。

+++ SimDisp/Remote/SimClient.cpp
@@ -66,7 +66,7 @@ class SiDiHost {
 					break;
 				case SIDI_MSG::SetOnKey:
 					if (lpfnOnKey)
-						lpfnOnKey(msg.ParamW(), msg.ParamL());
+						lpfnOnKey((uint16_t)msg.ParamW(), (uint8_t)msg.ParamL());
 					eventReq.Post(msg);
 					break;
 				case SIDI_MSG::SetOnResize:
